+++ scripts/src/admin/FormAdminPage.tsx
@@ -145,12 +145,18 @@ class FormAdminPage extends React.Component<IFormAdminPageProps, IFormAdminPageS
         }
         return (<Router>
             <div>
-            <Route path="/" render={(props) => <CenterList {...props} onError={e => this.onUnauth(e)} />} />
+            <Route path="/" render={(props) =>
+                <CenterList {...props} onError={e => this.onUnauth(e)} />}
+            />
             <Route path="/:centerSlug/:centerId" render={props =>
                 <FormList key={props.match.params.centerSlug} onError={e => this.onError(e, true)} userId={this.state.userId} {...props} />
             }/>
-            <Route path="/:centerSlug/:centerId/:formId/responses" render={props => <ResponseTable key={props.match.params.formId} onError={e => this.onError(e)} {...props} /> }/>
-            <Route path="/:centerSlug/:centerId/:formId/summary" render={props => <ResponseSummary key={props.match.params.formId} onError={e => this.onError(e)} {...props} /> }/>
+            <Route path="/:centerSlug/:centerId/:formId/responses" render={props =>
+                <ResponseTable key={props.match.params.formId} onError={e => this.onError(e)} {...props} />
+            }/>
+            <Route path="/:centerSlug/:centerId/:formId/summary" render={props =>
+                <ResponseSummary key={props.match.params.formId} onError={e => this.onError(e)} {...props} />
+            }/>
             </div>
         </Router>);