+++ CHANGELOG.md
@@ -37,7 +37,8 @@ Todo:
 
 ## 1.2.4 (TBD)
 - webpack minify css and put in a separate file
-- Highlight responses / summary buttons so you know what page you're on.
+- Highlight responses / summary buttons so you know what page you're on, and grey out buttons you have no permissions to.
+- show cff:cognitoIdentityId: before id
 
 ## 1.2.3 (4/2/2018)
 - Refactor loading, proper error handling for responses page